move record templates from src to package root
