+++ packages/record/src/recorder.spec.ts
@@ -7,7 +7,7 @@ import {
   RecorderInitializeError,
 } from "./error";
 import Recorder from "./recorder";
-import webcamFeed from "./templates/webcam-feed.mustache";
+import webcamFeed from "../templates/webcam-feed.mustache";
 import { CSSWidthHeight } from "./types";
 
 jest.mock("@lookit/data");
