+++ packages/record/src/stop.ts
@@ -3,7 +3,7 @@ import { JsPsych, JsPsychPlugin } from "jspsych";
 import Mustache from "mustache";
 import { NoSessionRecordingError } from "./error";
 import Recorder from "./recorder";
-import uploadingVideo from "./templates/uploading-video.mustache";
+import uploadingVideo from "../templates/uploading-video.mustache";
 
 declare let window: LookitWindow;
