+++ packages/record/templates/uploading-video.mustache
