+++ packages/record/templates/webcam-feed.mustache
